Refactor : 졸업시험 스크래핑, 스크래핑 데이터 파싱 코드 병합
