--- src/main/java/wiseSWlife/wiseSWlife/service/graduation/convertScrapingInf/ConvertExamTable.java
@@ -1,9 +1,0 @@
-package wiseSWlife.wiseSWlife.service.graduation.convertScrapingInf;
-
-import wiseSWlife.wiseSWlife.model.graduation.ExamTable;
-
-import java.util.Map;
-
-public interface ConvertExamTable {
-    Map<String, Boolean> convert(ExamTable table);
-}
