+++ src/main/java/wiseSWlife/wiseSWlife/service/graduation/scrapingImpl/Exam.java
@@ -9,6 +9,9 @@ import wiseSWlife.wiseSWlife.model.graduation.ExamTable;
 import wiseSWlife.wiseSWlife.service.graduation.scrapingInterface.ExamScraping;
 
 import java.io.*;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.Map;
 
 @Service
 @RequiredArgsConstructor
@@ -23,7 +26,7 @@ public class Exam implements ExamScraping {
     String pythonEncoding;
 
     @Override
-    public ExamTable scrapping(String intCookie) throws IOException, InterruptedException {
+    public ExamTable scraping(String intCookie) throws IOException, InterruptedException {
         File testFile = new File(pythonFile);
         FileWriter fw = new FileWriter(testFile);
         fw.write("import asyncio\n" + "\n");
@@ -53,5 +56,22 @@ public class Exam implements ExamScraping {
         ExamTable examTable = gson.fromJson(testLine,ExamTable.class);
 
         return examTable;
+    }
+
+    @Override
+    public Map<String, Boolean> convert(ExamTable table) {
+        Map<String, Boolean> examMap = new HashMap<>();
+        String[] subjects = {"성경", "영어", "컴퓨터", "컴퓨터2"};
+
+        for(String i : subjects){
+            examMap.put(i, false);
+        }
+
+        for(ArrayList i:table.getBody()){
+            if(i.get(4).equals("합격")){
+                examMap.put(i.get(1).toString(),true);
+            }
+        }
+        return examMap;
     }
 }
