+++ src/main/java/wiseSWlife/wiseSWlife/service/graduation/scrapingInterface/ExamScraping.java
@@ -3,8 +3,11 @@ package wiseSWlife.wiseSWlife.service.graduation.scrapingInterface;
 import wiseSWlife.wiseSWlife.model.graduation.ExamTable;
 
 import java.io.IOException;
+import java.util.Map;
 
 public interface ExamScraping {
 
-    ExamTable scrapping(String intCookie) throws IOException, InterruptedException;
+    ExamTable scraping(String intCookie) throws IOException, InterruptedException;
+
+    Map<String, Boolean> convert(ExamTable table);
 }
